+++ release_notes.md
@@ -1,4 +1,6 @@
 # Release Notes
 
 ## v0.1.0
+ * Created a base library to have some standard functions and definitions
  * Logging system and assertions in Fracture.
+ * Created a platform layer library to abstract platform specific commands.